+++ markdown-mode.el
@@ -1631,6 +1631,13 @@ or
   "Regular expression for matching inline identifiers or attribute lists.
 Compatible with Pandoc, Python Markdown, PHP Markdown Extra, and Leanpub.")
 
+(defconst markdown-regex-leanpub-sections
+  (concat
+   "^\\({\\)\\("
+   (regexp-opt '("frontmatter" "mainmatter" "backmatter" "appendix" "pagebreak"))
+   "\\)\\(}\\)[ \t]*\n")
+  "Regular expression for Leanpub section markers and related syntax.")
+
 (defconst markdown-regex-sub-superscript
   "\\(?:^\\|[^\\~^]\\)\\(\\([~^]\\)\\([[:alnum:]]+\\)\\(\\2\\)\\)"
   "The regular expression matching a sub- or superscript.
@@ -2654,6 +2661,7 @@ Depending on your font, some reasonable choices are:
     (,markdown-regex-line-break . (1 markdown-line-break-face prepend))
     (markdown-fontify-sub-superscripts)
     (markdown-match-inline-attributes . ((0 markdown-markup-properties prepend)))
+    (markdown-match-leanpub-sections . ((0 markdown-markup-properties)))
     (markdown-fontify-blockquotes))
   "Syntax highlighting for Markdown files.")
 
@@ -3680,6 +3688,14 @@ is \"\n\n\""
 (defun markdown-match-inline-attributes (last)
   "Match inline attributes from point to LAST."
   (when (markdown-match-inline-generic markdown-regex-inline-attributes last)
+    (unless (or (markdown-inline-code-at-pos-p (match-beginning 0))
+                (markdown-inline-code-at-pos-p (match-end 0))
+                (markdown-in-comment-p))
+      t)))
+
+(defun markdown-match-leanpub-sections (last)
+  "Match Leanpub section markers from point to LAST."
+  (when (markdown-match-inline-generic markdown-regex-leanpub-sections last)
     (unless (or (markdown-inline-code-at-pos-p (match-beginning 0))
                 (markdown-inline-code-at-pos-p (match-end 0))
                 (markdown-in-comment-p))